always use uint8 as type

+++ src/istore.h
@@ -98,8 +98,8 @@ struct ISPairs {
 
 typedef struct ISPairs ISPairs;
 
-extern void is_pairs_init(ISPairs *pairs, size_t initial_size, int type);
-extern void is_pairs_insert(ISPairs *pairs, int32 key, long val, int type);
+extern void is_pairs_init(ISPairs *pairs, size_t initial_size, uint8 type);
+extern void is_pairs_insert(ISPairs *pairs, int32 key, long val, uint8 type);
 extern int  is_pairs_cmp(const void *a, const void *b);
 extern void is_pairs_sort(ISPairs *pairs);
 extern void is_pairs_deinit(ISPairs *pairs);
